requirements.txt
requests==2.25.1
Pillow==8.0.1
telebot==0.0.4
